requirement.txt
streamlit
pandas
numpy
plotly.express
matplotlib
fastai==1.0.61
PIL
sqlalchemy
wheel
sk learn